fix: change value type to float

+++ Assets/Scripts/StatSystem/StatDataSO.cs
@@ -16,10 +16,10 @@ namespace Juhyeon.StatSystem
         #endregion
 
         #region Data
-        [Range(0, 100)] public byte HP;
-        [Range(0, 100)] public byte AP;
-        [Range(0, 100)] public byte MOV;
-        [Range(0, 100)] public byte ATK;
+        [Range(0, 100)] public float HP;
+        [Range(0, 100)] public float AP;
+        [Range(0, 100)] public float MOV;
+        [Range(0, 100)] public float ATK;
         #endregion
 
         public GameObject prefab;
